+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy
+opencv-python
+PIL
+scipy
+skimage
